refactor(updating): refactor for jspm
